Remove the need for react-screenshot-test.config.js
This should solve #43.
