+++ jest.screenshot.config.js
@@ -7,14 +7,13 @@ module.exports = {
   },
   transform: {
     "^.+\\.[t|j]sx?$": "ts-jest",
+    "^.+\\.module\\.css$": "./css-modules-transform",
     "^.+\\.css$": "./css-transform",
     "^.+\\.scss$": "./sass-transform",
     "^.+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$":
       "./asset-transform"
   },
-  transformIgnorePatterns: [
-    "node_modules/.+\\.js"
-  ],
+  transformIgnorePatterns: ["node_modules/.+\\.js"],
   testMatch: ["**/?(*.)+(screenshot).[jt]s?(x)"],
   testPathIgnorePatterns: ["/dist/", "/example/", "/node_modules/"]
 };
