--- react-screenshot-test.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  cssModules: name => name.endsWith(".module.css")
-};
